加了计算学分总分功能
还改了表score，course，add_teacher但是还没有做到功能实现

+++ stu_course.h
@@ -5,7 +5,7 @@
 #include <QStandardItem>
 #include <QStandardItemModel>
 
-extern float sum;
+extern int sum;
 
 namespace Ui {
 class stu_course;
